+++ src/middleware/jira-symmetric-jwt-middleware.test.ts
@@ -2,19 +2,18 @@ import { encodeSymmetric } from "atlassian-jwt";
 import { when } from "jest-when";
 import { booleanFlag, BooleanFlags } from "~/src/config/feature-flags";
 import { getLogger } from "~/src/config/logger";
-import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jiraSymmetricJwtMiddleware";
+import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 import { Installation } from "~/src/models/installation";
 
 
 const logger = getLogger("jira-jwt-verify-middleware.test");
-jest.mock("models/installation");
 jest.mock("config/feature-flags");
 const testSharedSecret = "test-secret";
 
 describe("jiraSymmetricJwtMiddleware", () => {
 	let res;
 	let next;
-	let installation;
+	//	let installation;
 
 	beforeEach(async () => {
 		res = {
@@ -27,27 +26,19 @@ describe("jiraSymmetricJwtMiddleware", () => {
 		res.status.mockReturnValue(res);
 		next = jest.fn();
 
-		installation = {
+		/* 	installation = {
 			id: 19,
 			jiraHost,
 			clientKey: "jira-client-key",
 			enabled: true,
 			decrypt: jest.fn(() => testSharedSecret),
 			subscriptions: jest.fn().mockResolvedValue([])
-		};
+		}; */
 
-		jest.mocked(Installation.getForClientKey).mockImplementation((clientKey) => {
-			if (clientKey == installation.clientKey)	{
-				return installation;
-			}
-			return null;
-		});
-
-		jest.mocked(Installation.getForHost).mockImplementation((host) => {
-			if (host == installation.jiraHost)	{
-				return installation;
-			}
-			return null;
+		await Installation.install({
+			clientKey: "jira-client-key",
+			host: jiraHost,
+			sharedSecret: testSharedSecret
 		});
 
 		when(booleanFlag).calledWith(
@@ -103,9 +94,9 @@ describe("jiraSymmetricJwtMiddleware", () => {
 		await jiraSymmetricJwtMiddleware(req, res, next);
 
 		expect(next).toHaveBeenCalledTimes(1);
-		expect(res.locals.installation).toEqual(installation);
-		expect(res.locals.jiraHost).toEqual(installation.jiraHost);
-		expect(req.session.jiraHost).toEqual(installation.jiraHost);
+		expect(res.locals.installation.jiraHost).toEqual(jiraHost);
+		expect(res.locals.jiraHost).toEqual(jiraHost);
+		expect(req.session.jiraHost).toEqual(jiraHost);
 	});
 
 	it("should throw error when token is wrong", async () => {
@@ -120,7 +111,7 @@ describe("jiraSymmetricJwtMiddleware", () => {
 
 	it("should call next when jiraHost set in session", async () => {
 		const req = buildRequestWithNoToken();
-		req.session.jiraHost = installation.jiraHost;
+		req.session.jiraHost = jiraHost;
 
 		await jiraSymmetricJwtMiddleware(req, res, next);
 
@@ -129,13 +120,13 @@ describe("jiraSymmetricJwtMiddleware", () => {
 
 	it("should set res.locals and req.session when jiraHost set in session", async () => {
 		const req = buildRequestWithNoToken();
-		req.session.jiraHost = installation.jiraHost;
+		req.session.jiraHost = jiraHost;
 
 		await jiraSymmetricJwtMiddleware(req, res, next);
 
 		expect(next).toHaveBeenCalledTimes(1);
-		expect(res.locals.installation).toEqual(installation);
-		expect(res.locals.jiraHost).toEqual(installation.jiraHost);
+		expect(res.locals.installation.jiraHost).toEqual(jiraHost);
+		expect(res.locals.jiraHost).toEqual(jiraHost);
 	});
 
 });
