+++ src/middleware/jira-symmetric-jwt-middleware.ts
@@ -1,4 +1,5 @@
 import { decodeSymmetric, getAlgorithm } from "atlassian-jwt";
+import Logger from "bunyan";
 import { NextFunction, Request, Response } from "express";
 import { booleanFlag, BooleanFlags } from "~/src/config/feature-flags";
 import { TokenType, verifyQsh } from "~/src/jira/util/jwt";
@@ -15,15 +16,15 @@ export const jiraSymmetricJwtMiddleware = async (req: Request, res: Response, ne
 	const token = req.query?.["jwt"] || req.cookies?.["jwt"];
 
 	if (token) {
-		const issuer = getIssuer(token);
+		const issuer = getIssuer(token, req.log);
 		if (!issuer) {
-			req.log.info("JWT claim did not contain the issuer (iss) claim");
+			req.log.warn("JWT claim did not contain the issuer (iss) claim");
 			return res.status(401).send("Unauthorised");
 		}
 
 		const installation = await Installation.getForClientKey(issuer);
 		if (!installation) {
-			req.log.info("No Installation found");
+			req.log.warn("No Installation found");
 			return res.status(401).send("Unauthorised");
 		}
 
@@ -32,9 +33,9 @@ export const jiraSymmetricJwtMiddleware = async (req: Request, res: Response, ne
 		const path = req.baseUrl == "" ? req.path : req.baseUrl;
 		const tokenType = checkPathValidity(path) ? TokenType.normal : TokenType.context;
 		try {
-			verifySymmetricJwt(token, secret, req, tokenType);
+			verifySymmetricJwt(token, secret, req, tokenType, req.log);
 		} catch (err) {
-			req.log.info(err.message);
+			req.log.warn({ err }, err.message);
 			return res.status(401).send("Unauthorised");
 		}
 
@@ -51,7 +52,8 @@ export const jiraSymmetricJwtMiddleware = async (req: Request, res: Response, ne
 
 		const installation = await Installation.getForHost(req.session.jiraHost);
 		if (!installation) {
-			throw new Error("No Installation found");
+			req.log.warn("No Installation found");
+			return res.status(401).send("Unauthorised");
 		}
 
 		res.locals.installation = installation;
@@ -64,40 +66,29 @@ export const jiraSymmetricJwtMiddleware = async (req: Request, res: Response, ne
 
 };
 
-const getIssuer = (token: string): string | undefined => {
+const getIssuer = (token: string, logger: Logger): string | undefined => {
 
 	let unverifiedClaims;
 	try {
 		unverifiedClaims = decodeSymmetric(token, "", getAlgorithm(token), true); // decode without verification;
-	} catch (e) {
-		throw new Error(`Invalid JWT: ${e.message}`);
+	} catch (err) {
+		logger.warn({ err }, "Invalid JWT");
+		throw new Error(`Invalid JWT: ${err.message}`);
 	}
 
 	return unverifiedClaims.iss;
 };
 
-const verifySymmetricJwt = (token: string, secret: string, req: Request, tokenType: TokenType): boolean => {
+const verifySymmetricJwt = (token: string, secret: string, req: Request, tokenType: TokenType, logger: Logger): boolean => {
 	const algorithm = getAlgorithm(token);
-
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	let unverifiedClaims: any;
-	try {
-		unverifiedClaims = decodeSymmetric(token, "", algorithm, true); // decode without verification;
-	} catch (e) {
-		throw new Error(`Invalid JWT: ${e.message}`);
-	}
-
-	const issuer = unverifiedClaims.iss;
-	if (!issuer) {
-		throw new Error("JWT claim did not contain the issuer (iss) claim");
-	}
 
 	/* eslint-disable @typescript-eslint/no-explicit-any*/
 	let verifiedClaims: any; //due to decodeSymmetric return any
 	try {
 		verifiedClaims = decodeSymmetric(token, secret, algorithm, false);
-	} catch (error) {
-		throw new Error(`Unable to decode JWT token: ${error}`);
+	} catch (err) {
+		logger.warn({ err }, "Invalid JWT");
+		throw new Error(`Unable to decode JWT token: ${err}`);
 	}
 
 	return verifyJwtClaims(verifiedClaims, tokenType, req);
