+++ src/routes/jira/connect/jira-connect-router.ts
@@ -3,7 +3,7 @@ import { csrfMiddleware } from "middleware/csrf-middleware";
 import { JiraConnectGet } from "./jira-connect-get";
 import { JiraConnectEnterpriseRouter } from "./enterprise/jira-connect-enterprise-router";
 import { JiraJwtTokenMiddleware } from "~/src/middleware/jira-jwt-middleware";
-import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jiraSymmetricJwtMiddleware";
+import { jiraSymmetricJwtMiddleware } from "~/src/middleware/jira-symmetric-jwt-middleware";
 
 export const JiraConnectRouter = Router();
